Rename analyzer.rb -> analysis.rb
