+++ lib/how_is/analysis.rb
